product microservice 0.0.4

+++ src/main/java/com/project/productMicroservice/dto/ProductDetailsOfMerchantDto.java
@@ -14,5 +14,5 @@ public class ProductDetailsOfMerchantDto {
     private String imageUrl;
     private Map<String ,String> productAttributes;
     private String productDescription;
-    private int productRating;
+    private float productRating;
 }
